+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - the **[Joplin](https://joplinapp.org/) note import** was improved
   (for [#2189](https://github.com/pbek/QOwnNotes/issues/2189))
     - the whole notebook / folder structure can now now also be imported as note sub-folders
+- added more Italian, Brazilian Portuguese translation (thank you rawfreeamy, i****myself)
 
 ## 21.7.9
 - the **[Joplin](https://joplinapp.org/) note import** was improved