+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # QOwnNotes Changelog
 
-## 21.7.10
+## 21.7.11
 - the **[Joplin](https://joplinapp.org/) note import** was improved
   (for [#2189](https://github.com/pbek/QOwnNotes/issues/2189))
-    - the whole notebook / folder structure can now now also be imported as note sub-folders
+    - if Joplin didn't set the the file extension of attachments or images in the export
+      we now will try to get it from the original file name
+    - importing of images from image html tags like `<img src=":/ID"/>` is now also supported
+    - the metadata of notes can now optionally be imported as frontmatter
+
+## 21.7.10
+    - the **[Joplin](https://joplinapp.org/) note import** was improved
+      (for [#2189](https://github.com/pbek/QOwnNotes/issues/2189))
+        - the whole notebook / folder structure can now now also be imported as note sub-folders
 - added more Italian, Brazilian Portuguese translation (thank you rawfreeamy, i****myself)
 
 ## 21.7.9